Finish Backend

+++ 01-class-1/controllers/tareasControllers.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler')
 const Tarea = require('../model/tareaModel')
 
 const getTareas = asyncHandler(async (req, res) => {
-  const tareas = await Tarea.find()
+  const tareas = await Tarea.find({ user: req.user.id })
   res.status(200).json(tareas)
 })
 
@@ -15,7 +15,7 @@ const crearTareas = asyncHandler(async (req, res) => {
 
   const tarea = await Tarea.create({
     texto: req.body.texto,
-    user: req.body.user
+    user: req.user.id
   })
 
   res.status(201).json(tarea)
@@ -29,21 +29,34 @@ const updateTareas = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('tarea no encontrada')
   } else {
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
-    res.status(200).json(tareaUpdated)
+
+    // verificar que la tarea que queremos modificar pertenece al usuario logeado
+    if (tarea.user.toString() !== req.user.id) {
+      res.status(401)
+      throw new Error('Usuario no autorizado')
+    } else {
+      const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+      res.status(200).json(tareaUpdated)
+    }
   }
 
 })
 
 const deleteTareas = asyncHandler(async (req, res) => {
+
   const tarea = await Tarea.findById(req.params.id)
 
   if (!tarea) {
     res.status(400)
     throw new Error('Tarea no encontrada')
   } else {
-    await tarea.deleteOne()
-    res.status(200).json({ id: req.params.id })
+    if (tarea.user.toString() !== req.user.id) {
+      res.status(401)
+      throw new Error('Usuario no autorizado')
+    } else {
+      await tarea.deleteOne()
+      res.status(200).json({ id: req.params.id })
+    }
   }
 })
